# docs/openapi/bookkin-api.yaml
openapi: 3.1.0
info:
  title: BookKin API
  version: 1.0.0
  description: 家庭 NAS 电子图书馆的浏览、阅读、用户与受控文件事务 API。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /auth/setup:
    post:
      tags: [Auth]
      security: []
      operationId: setupOwner
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SetupRequest' } } } }
      responses:
        '200': { description: Owner session, content: { application/json: { schema: { $ref: '#/components/schemas/SessionUser' } } } }
        '409': { $ref: '#/components/responses/Problem' }
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      operationId: login
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses:
        '200': { description: Authenticated session, content: { application/json: { schema: { $ref: '#/components/schemas/SessionUser' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /auth/password:
    put:
      tags: [Auth]
      operationId: changePassword
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ChangePasswordRequest' } } } }
      responses:
        '200': { description: Updated session, content: { application/json: { schema: { $ref: '#/components/schemas/SessionUser' } } } }
...
  /booklists:
    get:
      tags: [Booklists]
      security: []
      operationId: listBooklists
      parameters: [{ name: q, in: query, schema: { type: string } }]
      responses:
        '200': { description: Booklists visible to the current audience, content: { application/json: { schema: { $ref: '#/components/schemas/BooklistList' } } } }
    post:
      tags: [Booklists]
      operationId: createBooklist
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateBooklistRequest' } } } }
      responses:
        '200': { description: Created booklist, content: { application/json: { schema: { $ref: '#/components/schemas/BooklistDetail' } } } }
  /booklists/{id}:
    get:
      tags: [Booklists]
      security: []
      operationId: getBooklist
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Visible booklist, content: { application/json: { schema: { $ref: '#/components/schemas/BooklistDetail' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Booklists]
      operationId: updateBooklist
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateBooklistRequest' } } } }
      responses:
        '200': { description: Updated booklist, content: { application/json: { schema: { $ref: '#/components/schemas/BooklistDetail' } } } }
        '409': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Booklists]
      operationId: deleteBooklist
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: revision, in: query, required: true, schema: { type: integer, format: int64, minimum: 0 } }
      responses:
        '204': { description: Booklist relationship removed without deleting books }
        '409': { $ref: '#/components/responses/Problem' }
  /booklists/{id}/books:
    get:
      tags: [Booklists]
      security: []
      operationId: listBooklistBooks
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: q, in: query, schema: { type: string } }
        - { name: format, in: query, schema: { $ref: '#/components/schemas/BookFormat' } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 36 } }
      responses:
        '200': { description: Visible books in manual list order, content: { application/json: { schema: { $ref: '#/components/schemas/BrowseBookPage' } } } }
    post:
      tags: [Booklists]
      operationId: addBooklistBooks
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AddBooklistBooksRequest' } } } }
      responses:
        '200': { description: Updated booklist, content: { application/json: { schema: { $ref: '#/components/schemas/BooklistDetail' } } } }
        '409': { $ref: '#/components/responses/Problem' }
  /booklists/{id}/books/{bookId}:
    delete:
      tags: [Booklists]
      operationId: removeBooklistBook
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: bookId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: revision, in: query, required: true, schema: { type: integer, format: int64, minimum: 0 } }
      responses:
        '200': { description: Updated booklist, content: { application/json: { schema: { $ref: '#/components/schemas/BooklistDetail' } } } }
  /booklists/{id}/order:
    put:
      tags: [Booklists]
      operationId: reorderBooklistBooks
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReorderBooklistRequest' } } } }
      responses:
        '200': { description: Reordered booklist, content: { application/json: { schema: { $ref: '#/components/schemas/BooklistDetail' } } } }
        '409': { $ref: '#/components/responses/Problem' }
...
  /books/{id}:
    get:
      tags: [Catalog]
      operationId: getBook
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Book, content: { application/json: { schema: { $ref: '#/components/schemas/Book' } } } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: BOOKKIN_SESSION
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Problem:
      description: RFC 9457 problem
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Problem:
      type: object
      properties: { type: { type: string, format: uri }, title: { type: string }, status: { type: integer }, detail: { type: string }, code: { type: string } }
...
    UserRole: { type: string, enum: [OWNER, ADMIN, MEMBER] }
...
    SessionUser:
      type: object
      required: [id, username, displayName, role, mustChangePassword]
      properties: { id: { type: string, format: uuid }, username: { type: string }, displayName: { type: string }, role: { $ref: '#/components/schemas/UserRole' }, mustChangePassword: { type: boolean } }
...
    SetupRequest:
      type: object
      required: [username, displayName, password]
      properties: { username: { type: string, pattern: '^[a-zA-Z0-9._-]{3,80}$' }, displayName: { type: string, maxLength: 120 }, password: { type: string, minLength: 12, maxLength: 200 } }
...
    LoginRequest: { type: object, required: [username, password], properties: { username: { type: string }, password: { type: string } } }
    ChangePasswordRequest: { type: object, required: [currentPassword, newPassword], properties: { currentPassword: { type: string }, newPassword: { type: string, minLength: 12 } } }
...
    BookFormat: { type: string, enum: [EPUB, PDF] }
    BookFileStatus: { type: string, enum: [AVAILABLE, OPERATING, TRASHED, MISSING, DELETED] }
...
    Book:
      type: object
      required: [id, fileId, title, author, format, coverUrl, progress, addedAt, libraryRoot, relativePath, fingerprint, status, tags]
      properties:
        id: { type: string, format: uuid }
        fileId: { type: string, format: uuid }
        title: { type: string }
        author: { type: string }
        series: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        format: { $ref: '#/components/schemas/BookFormat' }
        coverUrl: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
        wordCount: { type: [string, 'null'] }
        addedAt: { type: string, format: date-time }
        libraryRoot: { type: string }
        relativePath: { type: string }
        fingerprint: { type: string }
        status: { $ref: '#/components/schemas/BookFileStatus' }
        tags: { type: array, items: { type: string } }
...
    BrowseBook:
      type: object
      required: [id, title, author, coverUrl, available, addedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        author: { type: string }
        description: { type: [string, 'null'] }
        format: { oneOf: [{ $ref: '#/components/schemas/BookFormat' }, { type: 'null' }] }
        coverUrl: { type: string }
        available: { type: boolean }
        addedAt: { type: string, format: date-time }
    BrowseBookPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/BrowseBook' } }
        nextCursor: { type: [string, 'null'] }
...
    BooklistKind: { type: string, enum: [OFFICIAL, PERSONAL] }
    BooklistVisibility: { type: string, enum: [PRIVATE, MEMBERS, PUBLIC] }
    BooklistSummary:
      type: object
      required: [id, title, kind, visibility, ownerDisplayName, bookCount, previewBooks, ownedByViewer, editable, revision, hiddenPublicBookCount, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: [string, 'null'] }
        kind: { $ref: '#/components/schemas/BooklistKind' }
        visibility: { $ref: '#/components/schemas/BooklistVisibility' }
        ownerDisplayName: { type: string }
        bookCount: { type: integer, format: int64, minimum: 0 }
        previewBooks: { type: array, maxItems: 4, items: { $ref: '#/components/schemas/BrowseBook' } }
        ownedByViewer: { type: boolean }
        editable: { type: boolean }
        revision: { type: integer, format: int64, minimum: 0 }
        hiddenPublicBookCount: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
    BooklistDetail:
      allOf:
        - { $ref: '#/components/schemas/BooklistSummary' }
    BooklistList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/BooklistSummary' } } }
    CreateBooklistRequest:
      type: object
      required: [title, kind, visibility]
      properties:
        title: { type: string, maxLength: 240 }
        description: { type: [string, 'null'], maxLength: 4000 }
        kind: { $ref: '#/components/schemas/BooklistKind' }
        visibility: { $ref: '#/components/schemas/BooklistVisibility' }
    UpdateBooklistRequest:
      type: object
      required: [title, visibility, revision]
      properties:
        title: { type: string, maxLength: 240 }
        description: { type: [string, 'null'], maxLength: 4000 }
        visibility: { $ref: '#/components/schemas/BooklistVisibility' }
        revision: { type: integer, format: int64, minimum: 0 }
    AddBooklistBooksRequest:
      type: object
      required: [bookIds, revision]
      properties:
        bookIds: { type: array, items: { type: string, format: uuid } }
        revision: { type: integer, format: int64, minimum: 0 }
    ReorderBooklistRequest:
      type: object
      required: [bookIds, revision]
      properties:
        bookIds: { type: array, items: { type: string, format: uuid } }
        revision: { type: integer, format: int64, minimum: 0 }
...